# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookVibe API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Book {
  id: number;
  external_id?: string;
  source?: string;
  title: string;
  author?: string;
  description?: string;
  cover?: string;
  pages?: number;
  read_pages?: number;
  isbn?: string;
  status: string;
  shelf: string;
  vibe: string[];
  rating?: number;
  created_at?: string;
}

export interface BookSearchItem {
  external_id: string;
  source: string;
  title: string;
  author?: string;
  description?: string;
  cover?: string;
  pages?: number;
  isbn?: string;
  genres?: string[];
}

export interface BookSearchResult {
  query?: string;
  items: BookSearchItem[];
}

export interface BookListResult {
  items: Book[];
}

export interface BookItemResult {
  item: Book;
}

export interface SaveBookBody {
  external_id?: string;
  source?: string;
  title: string;
  author?: string;
  description?: string;
  cover?: string;
  pages?: number;
  read_pages?: number;
  isbn?: string;
  status: string;
  shelf: string;
  vibe?: string[];
  rating?: number;
}

export interface DeleteResult {
  deleted: boolean;
  id?: number;
}

export type DiaryEntryRatings = { [key: string]: number };

export type DiaryEntryMusic = { [key: string]: string };

export interface DiaryEntry {
  id: number;
  book_id: number;
  quote?: string;
  note?: string;
  ratings?: DiaryEntryRatings;
  music?: DiaryEntryMusic;
  images?: string[];
  stickers?: string[];
  created_at?: string;
  updated_at?: string;
}

export interface DiaryListResult {
  items: DiaryEntry[];
}

export interface DiaryItemResult {
  item: DiaryEntry;
}

export type SaveDiaryBodyRatings = { [key: string]: number };

export type SaveDiaryBodyMusic = { [key: string]: string };

export interface SaveDiaryBody {
  book_id: number;
  quote?: string;
  note?: string;
  ratings?: SaveDiaryBodyRatings;
  music?: SaveDiaryBodyMusic;
  images?: string[];
  stickers?: string[];
}

export interface ChatMessage {
  id: number;
  room_id: string;
  author: string;
  text: string;
  reply_to?: number | null;
  sticker?: string;
  image_url?: string;
  created_at?: string;
}

export interface ChatRoom {
  id: string;
  name: string;
  description?: string;
  messages?: number;
  last_message?: ChatMessage;
}

export interface ChatRoomListResult {
  items: ChatRoom[];
}

export interface ChatMessageListResult {
  items: ChatMessage[];
}

export interface ChatMessageItemResult {
  item: ChatMessage;
}

export interface SendMessageBody {
  author: string;
  text?: string;
  reply_to?: number | null;
  sticker?: string;
  image_url?: string;
}

export interface MusicTrack {
  title: string;
  artist: string;
  album?: string;
  cover?: string;
  preview?: string;
  link?: string;
  source?: string;
}

export interface MusicSearchResult {
  query?: string;
  items: MusicTrack[];
}

export interface ImageItem {
  title?: string;
  url: string;
  thumb?: string;
  author?: string;
  source?: string;
}

export interface ImageSearchResult {
  query?: string;
  items: ImageItem[];
}

export interface AiDiaryHelperBody {
  book_title: string;
  mood?: string;
  note?: string;
}

export interface AiDiaryHelperResult {
  source: string;
  result: string;
}

export type SearchBooksParams = {
  q: string;
  limit?: number;
};

export type ListChatMessagesParams = {
  limit?: number;
};

export type SearchMusicParams = {
  q: string;
  limit?: number;
};

export type SearchImagesParams = {
  q: string;
  limit?: number;
};