# lib/api-zod/src/generated/types/bookSearchItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookVibe API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BookSearchItem {
  external_id: string;
  source: string;
  title: string;
  author?: string;
  description?: string;
  cover?: string;
  pages?: number;
  isbn?: string;
  genres?: string[];
}