# lib/api-zod/src/generated/types/chatRoomListResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookVibe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatRoom } from "./chatRoom";

export interface ChatRoomListResult {
  items: ChatRoom[];
}