# lib/api-zod/src/generated/types/deleteResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookVibe API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DeleteResult {
  deleted: boolean;
  id?: number;
}