# lib/api-zod/src/generated/types/musicSearchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookVibe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MusicTrack } from "./musicTrack";

export interface MusicSearchResult {
  query?: string;
  items: MusicTrack[];
}